+++ src/app.java
@@ -1,0 +1,10 @@
+public class app {
+    public static void main(String[] args) {
+        opcList o = new opcList();
+        o.Import();
+        ListaEmpleados listaEmpleados = o.getListaEmpleados();
+        Menu m = new Menu(listaEmpleados);
+        m.Login();
+        
+    }
+}